Version0.2

+++ Programa2.py
@@ -7,7 +7,7 @@ import numpy as np
 
 miLista = [[1,2,3,4],[5,6,7,8],[9,10,11,20]]
 
-miArreglo = np.array(miLista)
+miArreglo = np.array(miLista, dtype = int)
 print(miArreglo.ndim)
 print(miArreglo)
 print(miLista)
